requirements.txt
asgiref==3.5.2
boto==2.49.0
boto3==1.22.8
botocore==1.25.8
Django==4.0.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
jmespath==1.0.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
